Added v2 of script
Did a total rewrite/cleanup of the script. It now uses cookies instead
of logging in each poll. It now also updates everytime someone requests
the page. Normally this wouldn't be good, but now we run this locally
and know that it is only our rssreader that reads the value.